# docs/03.application-development/02.design/04.openapi/openapi.yaml
openapi: 3.1.0
info:
  title: BMS API
  version: 1.0.0
  description: Business Management System API contract
servers:
  - url: /api/v1
security:
  - sessionCookie: []
tags:
  - name: Authentication
    description: 로그인과 서버 세션 관리
  - name: UserManagement
    description: 내부 직원과 사용자 계정 관리
paths:
  /auth/login:
    post:
      tags: [Authentication]
      summary: 로그인
      operationId: commonLogin
      security:
        - csrfToken: []
      x-bms-api-id: API-COM-AUTH-001
      x-bms-requirement-ids: [REQ-COM-005, NFR-SEC-001, NFR-SEC-002]
      x-bms-function-ids: [BFD-02-05-01-01]
      x-bms-screen-ids: [COM-001]
      x-bms-permission: PUBLIC
      x-bms-error-codes: [COMMON_VALIDATION_FAILED, AUTH_LOGIN_FAILED, AUTH_CSRF_INVALID, AUTH_TOO_MANY_ATTEMPTS, COMMON_INTERNAL_ERROR]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 로그인 성공과 서버 세션 생성
          headers:
            Set-Cookie:
              description: BMSSESSION; Path=/; HttpOnly; 운영 Secure; SameSite=Lax; Domain·Max-Age·Expires 미지정
              schema:
                type: string
            Cache-Control:
              description: 인증 응답 캐시 금지
              schema:
                type: string
                example: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/LoginFailedProblem'
        '403':
          $ref: '#/components/responses/ForbiddenProblem'
        '429':
          $ref: '#/components/responses/TooManyAttemptsProblem'
        '500':
          $ref: '#/components/responses/InternalProblem'
  /auth/logout:
    post:
      tags: [Authentication]
      summary: 로그아웃
      operationId: commonLogout
      security:
        - sessionCookie: []
          csrfToken: []
      x-bms-api-id: API-COM-AUTH-002
      x-bms-requirement-ids: [REQ-COM-005]
      x-bms-function-ids: [BFD-02-05-01-02]
      x-bms-screen-ids: [COM-001]
      x-bms-permission: AUTHENTICATED_OR_LIMITED
      x-bms-error-codes: [AUTH_AUTHENTICATION_REQUIRED, AUTH_CSRF_INVALID, COMMON_INTERNAL_ERROR]
      responses:
        '204':
          description: 세션과 인증 쿠키 제거 완료
        '401':
          $ref: '#/components/responses/AuthenticationRequiredProblem'
        '403':
          $ref: '#/components/responses/ForbiddenProblem'
        '500':
          $ref: '#/components/responses/InternalProblem'
  /auth/me:
    get:
      tags: [Authentication]
      summary: 현재 사용자 조회
      operationId: commonGetCurrentUser
      security:
        - sessionCookie: []
      x-bms-api-id: API-COM-AUTH-003
      x-bms-requirement-ids: [REQ-COM-005, NFR-SEC-002]
      x-bms-function-ids: [BFD-02-05-01-03]
      x-bms-screen-ids: [COM-001]
      x-bms-permission: AUTHENTICATED_OR_LIMITED
      x-bms-error-codes: [AUTH_AUTHENTICATION_REQUIRED, COMMON_INTERNAL_ERROR]
      responses:
        '200':
          description: 로그인 시점의 사용자와 권한 스냅샷
          headers:
            Cache-Control:
              description: 인증 응답 캐시 금지
              schema:
                type: string
                example: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
        '401':
          $ref: '#/components/responses/AuthenticationRequiredProblem'
        '500':
          $ref: '#/components/responses/InternalProblem'
  /auth/activity:
    post:
      tags: [Authentication]
      summary: 사용자 활동시각 갱신
      operationId: commonRefreshActivity
      security:
        - sessionCookie: []
          csrfToken: []
      x-bms-api-id: API-COM-AUTH-004
      x-bms-requirement-ids: [REQ-COM-005]
      x-bms-function-ids: [BFD-02-05-01-03]
      x-bms-screen-ids: [COM-001]
      x-bms-permission: AUTHENTICATED_OR_LIMITED
      x-bms-error-codes: [AUTH_AUTHENTICATION_REQUIRED, AUTH_CSRF_INVALID, COMMON_INTERNAL_ERROR]
      responses:
        '204':
          description: 마지막 사용자 활동시각 갱신 완료
        '401':
          $ref: '#/components/responses/AuthenticationRequiredProblem'
        '403':
          $ref: '#/components/responses/ForbiddenProblem'
        '500':
          $ref: '#/components/responses/InternalProblem'
  /auth/reauthenticate:
    post:
      tags: [Authentication]
      summary: 중요 작업용 현재 비밀번호 재인증
      operationId: commonReauthenticate
      security:
        - sessionCookie: []
          csrfToken: []
      x-bms-api-id: API-COM-AUTH-006
      x-bms-requirement-ids: [REQ-COM-005, NFR-SEC-001, NFR-SEC-002]
      x-bms-function-ids: [BFD-02-05-01-01]
      x-bms-screen-ids: [COM-001]
      x-bms-permission: AUTHENTICATED
      x-bms-error-codes: [COMMON_VALIDATION_FAILED, AUTH_AUTHENTICATION_REQUIRED, AUTH_REAUTHENTICATION_FAILED, AUTH_CSRF_INVALID, AUTH_PASSWORD_CHANGE_REQUIRED, AUTH_TOO_MANY_ATTEMPTS, COMMON_INTERNAL_ERROR]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReauthenticateRequest'
      responses:
        '204':
          description: 현재 세션의 재인증시각 갱신 완료
          headers:
            Cache-Control:
              description: 재인증 응답 캐시 금지
              schema:
                type: string
                example: no-store
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthenticationRequiredProblem'
        '403':
          $ref: '#/components/responses/ForbiddenProblem'
        '429':
          $ref: '#/components/responses/TooManyAttemptsProblem'
        '500':
          $ref: '#/components/responses/InternalProblem'
  /auth/csrf:
    get:
      tags: [Authentication]
      summary: CSRF 토큰 조회
      operationId: commonGetCsrfToken
      security: []
      x-bms-api-id: API-COM-AUTH-005
      x-bms-requirement-ids: [NFR-SEC-001]
      x-bms-function-ids: [BFD-02-05-01-01]
      x-bms-screen-ids: [COM-001]
      x-bms-permission: PUBLIC
      x-bms-error-codes: [COMMON_INTERNAL_ERROR]
      responses:
        '200':
          description: 상태 변경 요청용 CSRF 토큰
          headers:
            Cache-Control:
              description: 토큰 응답 캐시 금지
              schema:
                type: string
                example: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CsrfTokenResponse'
        '500':
          $ref: '#/components/responses/InternalProblem'
  /users:
    get:
      tags: [UserManagement]
      summary: 사용자 목록 조회
      operationId: systemSearchUsers
      security:
        - sessionCookie: []
      x-bms-api-id: API-SYS-USER-001
      x-bms-requirement-ids: [REQ-SYS-001, NFR-SEC-003]
      x-bms-function-ids: [BFD-01-01-01-01]
      x-bms-screen-ids: [SYS-001]
      x-bms-permission: SYS-001
      x-bms-error-codes: [COMMON_VALIDATION_FAILED, AUTH_AUTHENTICATION_REQUIRED, AUTH_FORBIDDEN, AUTH_PASSWORD_CHANGE_REQUIRED, COMMON_INTERNAL_ERROR]
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
            maxLength: 100
        - name: accountStatusCode
          in: query
          schema:
            type: string
            enum: [ACTIVE, INACTIVE]
        - name: organizationId
          in: query
          schema:
            type: string
        - name: roleId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sort
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: 사용자 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSearchResponse'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthenticationRequiredProblem'
        '403':
          $ref: '#/components/responses/ForbiddenProblem'
        '500':
          $ref: '#/components/responses/InternalProblem'
    post:
      tags: [UserManagement]
      summary: 내부 직원과 사용자 계정 등록
      operationId: systemCreateUser
      security:
        - sessionCookie: []
          csrfToken: []
      x-bms-api-id: API-SYS-USER-002
      x-bms-requirement-ids: [REQ-SYS-001, REQ-RES-001, NFR-SEC-003]
      x-bms-function-ids: [BFD-01-01-01-02, BFD-07-01-01-03, BFD-07-01-02-02]
      x-bms-screen-ids: [SYS-001]
      x-bms-permission: SYS-001
      x-bms-error-codes: [COMMON_VALIDATION_FAILED, AUTH_AUTHENTICATION_REQUIRED, AUTH_FORBIDDEN, AUTH_CSRF_INVALID, AUTH_PASSWORD_CHANGE_REQUIRED, COMMON_RESOURCE_NOT_FOUND, SYS_LOGIN_ID_DUPLICATE, RES_EMPLOYEE_NUMBER_DUPLICATE, COMMON_INVALID_STATE, COMMON_INTERNAL_ERROR]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: 직원·사용자·초기 역할 등록 완료
          headers:
            Location:
              description: 생성된 사용자 자원 URL
              schema:
                type: string
                example: /api/v1/users/01K00000000000000000000000
            Cache-Control:
              description: 임시 비밀번호 응답 캐시 금지
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponse'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthenticationRequiredProblem'
        '403':
          $ref: '#/components/responses/ForbiddenProblem'
        '404':
          $ref: '#/components/responses/NotFoundProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '500':
          $ref: '#/components/responses/InternalProblem'
  /users/registration-options:
    get:
      tags: [UserManagement]
      summary: 사용자 등록 옵션 조회
      operationId: systemGetUserRegistrationOptions
      security:
        - sessionCookie: []
      x-bms-api-id: API-SYS-USER-003
      x-bms-requirement-ids: [REQ-SYS-001]
      x-bms-function-ids: [BFD-01-01-01-02]
      x-bms-screen-ids: [SYS-001]
      x-bms-permission: SYS-001
      x-bms-error-codes: [AUTH_AUTHENTICATION_REQUIRED, AUTH_FORBIDDEN, AUTH_PASSWORD_CHANGE_REQUIRED, COMMON_INTERNAL_ERROR]
      responses:
        '200':
          description: 등록 가능한 조직과 역할 옵션
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRegistrationOptionsResponse'
        '401':
          $ref: '#/components/responses/AuthenticationRequiredProblem'
        '403':
          $ref: '#/components/responses/ForbiddenProblem'
        '500':
          $ref: '#/components/responses/InternalProblem'
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: BMSSESSION
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
  schemas:
    UserSearchResponse:
      type: object
      additionalProperties: false
      required: [items, page, size, totalElements, totalPages, sort]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserSummary'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        sort:
          type: array
          items:
            type: string
    UserSummary:
      type: object
      additionalProperties: false
      required:
        - userId
        - employeeId
        - employeeNumber
        - loginId
        - userName
        - organizationId
        - organizationName
        - roleNames
        - accountStatusCode
        - passwordChangeRequired
        - registeredAt
      properties:
        userId:
          type: string
        employeeId:
          type: string
        employeeNumber:
          type: string
        loginId:
          type: string
        userName:
          type: string
        organizationId:
          type: string
        organizationName:
          type: string
        roleNames:
          type: array
          uniqueItems: true
          items:
            type: string
        emailMasked:
          type: [string, 'null']
        mobileMasked:
          type: [string, 'null']
        accountStatusCode:
          type: string
          enum: [ACTIVE, INACTIVE]
        passwordChangeRequired:
          type: boolean
        lastLoginAt:
          type: [string, 'null']
          format: date-time
        registeredAt:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      additionalProperties: false
      required: [employeeNumber, userName, organizationId, loginId, initialRoleIds]
      properties:
        employeeNumber:
          type: string
          minLength: 1
          maxLength: 30
        userName:
          type: string
          minLength: 1
          maxLength: 100
        organizationId:
          type: string
          minLength: 1
        loginId:
          type: string
          minLength: 4
          maxLength: 26
          pattern: '^[A-Za-z0-9][A-Za-z0-9._-]{3,25}$'
        initialRoleIds:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            minLength: 1
        email:
          type: string
          format: email
          maxLength: 100
        mobileNumber:
          type: string
          maxLength: 20
          pattern: '^[0-9-]+$'
    CreateUserResponse:
      type: object
      additionalProperties: false
      required: [userId, employeeId, loginId, temporaryPassword, temporaryPasswordExpiresAt]
      properties:
        userId:
          type: string
        employeeId:
          type: string
        loginId:
          type: string
        temporaryPassword:
          type: string
          format: password
          readOnly: true
          minLength: 16
          maxLength: 16
        temporaryPasswordExpiresAt:
          type: string
          format: date-time
          description: 서버에 저장한 최초 등록 전용 임시 비밀번호의 발급 후 24시간 만료일시
    UserRegistrationOptionsResponse:
      type: object
      additionalProperties: false
      required: [organizations, roles]
      properties:
        organizations:
          type: array
          items:
            $ref: '#/components/schemas/OrganizationOption'
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleOption'
    OrganizationOption:
      type: object
      additionalProperties: false
      required: [organizationId, organizationName]
      properties:
        organizationId:
          type: string
        organizationName:
          type: string
    RoleOption:
      type: object
      additionalProperties: false
      required: [roleId, roleName]
      properties:
        roleId:
          type: string
        roleName:
          type: string
    LoginRequest:
      type: object
      additionalProperties: false
      required: [loginId, password]
      properties:
        loginId:
          type: string
          minLength: 1
          maxLength: 100
          description: 앞뒤 공백 제거 후 영문자를 소문자로 변환하여 조회
        password:
          type: string
          format: password
          minLength: 1
          maxLength: 256
          writeOnly: true
    ReauthenticateRequest:
      type: object
      additionalProperties: false
      required: [password]
      properties:
        password:
          type: string
          format: password
          minLength: 1
          maxLength: 256
          writeOnly: true
    CurrentUserResponse:
      type: object
      additionalProperties: false
      required: [user, roles, menus, passwordChangeRequired, idleTimeoutSeconds, absoluteSessionExpiresAt]
      properties:
        user:
          $ref: '#/components/schemas/AuthenticatedUser'
        roles:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/AuthenticatedRole'
        menus:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/AuthorizedMenu'
        passwordChangeRequired:
          type: boolean
        idleTimeoutSeconds:
          type: integer
          minimum: 1
          example: 900
          description: 일반 세션 900초, 초기 비밀번호 제한 세션 600초
        absoluteSessionExpiresAt:
          type: string
          format: date-time
          description: 일반 세션은 로그인·승격 후 8시간, 제한 세션은 로그인 후 30분과 임시 비밀번호 만료 중 빠른 시각
    AuthenticatedUser:
      type: object
      additionalProperties: false
      required: [userId, loginId, displayName]
      properties:
        userId:
          type: string
        loginId:
          type: string
        displayName:
          type: string
    AuthenticatedRole:
      type: object
      additionalProperties: false
      required: [roleId, roleName]
      properties:
        roleId:
          type: string
        roleName:
          type: string
    AuthorizedMenu:
      type: object
      additionalProperties: false
      required: [menuId, menuName, sortOrder]
      properties:
        menuId:
          type: string
        parentMenuId:
          type: [string, 'null']
        menuName:
          type: string
        menuUrl:
          type: [string, 'null']
        sortOrder:
          type: integer
          minimum: 0
    CsrfTokenResponse:
      type: object
      additionalProperties: false
      required: [headerName, token]
      properties:
        headerName:
          type: string
          const: X-CSRF-TOKEN
        token:
          type: string
          minLength: 1
    FieldError:
      type: object
      required: [field, reason, message]
      properties:
        field:
          type: string
        reason:
          type: string
        message:
          type: string
    Problem:
      type: object
      required: [code, message, traceId]
      properties:
        code:
          type: string
        message:
          type: string
        traceId:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    PageMetadata:
      type: object
      required: [page, size, totalElements, totalPages, sort]
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        sort:
          type: array
          items:
            type: string
  responses:
    ValidationProblem:
      description: 요청값 검증 실패
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    LoginFailedProblem:
      description: 계정 존재 여부를 노출하지 않는 인증 실패
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    AuthenticationRequiredProblem:
      description: 인증 세션 없음 또는 만료
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFoundProblem:
      description: 조직·역할 또는 요청 자원을 찾을 수 없음
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ConflictProblem:
      description: 중복 자원 또는 현재 상태에서 처리할 수 없음
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ForbiddenProblem:
      description: CSRF 토큰 없음·불일치, 재인증 필요 또는 접근 거부
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    TooManyAttemptsProblem:
      description: 로그인 또는 재인증 요청 속도 제한 초과
      headers:
        Retry-After:
          description: 다시 요청할 수 있을 때까지의 초
          schema:
            type: integer
            minimum: 1
            example: 60
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    InternalProblem:
      description: 내부 서버 오류
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'